server test

+++ server.py
@@ -129,7 +129,7 @@ class AggregateCustomMetricStrategy(fl.server.strategy.FedAvgM):
         cur_global_params = parameters_to_ndarrays(self.initial_parameters)
         #test1 = lr_vector*weights_prime
         #test2 = cur_global_params + test1
-        new_global_params = (cur_global_params + weights_prime).float()
+        new_global_params = (cur_global_params + weights_prime)
         new_global_params = ndarrays_to_parameters(new_global_params)
 
         #metric stuff
